requirements.txt
django >= 3.1
requests >= 2.24.0
gunicorn >= 20.0.4
mysqlclient >= 2.0.1
python-dateutil >= 2.8.1
